Add first steps towards continuous Game of Life
Contains some classes to reduce the complexity of interacting with
GLFW/Opengl. The classes can be found in src/glfw-abstraction

A very basic version of Game Of Life has been implemented with both Compute and Fragment Shaders.
A benchmark script for them can be found in benchmark/game_of_life_simple. All the following
already implemented steps are implemented with Fragment shaders, since the benchmark
suggests that they are more effective for this kind of problem.

Adding onto the simple Game of Life, a kernel has been added. With this, an area with weights to
look at can be selected using a matrix. The kernel version still behaves exactly like the
normal Game of Life, since the kernel is selected so that a cell only looks at its
direct neighbours.

In the same way, a growth method was added to the Game of Life. This allows to add custom behaviour
on how the cell should grow / decay.

Using this the Larger Than Life version of the Game of Life has been implemented, the current
implementation initializes with a repeating and moving pattern called "Bosco".