Add continuous space
Adds continuous space (=^ bigger kernels). We can now bind kernels as uniforms and use those to get
vastly different behaviour.

It might be good idea to replace arrays with textures in the future, since the arrays stop
working after R=15 (961 floats in the array), which might be related to restrictions
on array sizes.