Fix bug when printing compile error information
When a shader compilation fails, the compile error as well as the shader code gets printed to cerr.

When printing the code, checking if the end of the code has been reached is equal to checking if it,
still contains line breaks. This is equal to position == -1. But since position was of type auto
which in this case is an unsigned type, the state position < 0 is unreachable, to fix this the
type of position can simply be changed to int.

This might cause errors if a single code line is longer than INT_MAX but this seems very
unlikely.